Add files via upload
Repository description:

This repository contains a Python script that can be used to automatically backup watched files from your Jellyfin server to a local or remote storage device.
The script works by first fetching a list of all the watched files from your Jellyfin server. It then uses this list to move the files to the specified backup location.
The script can be customized to backup specific libraries, seasons, or episodes. It can also be scheduled to run on a regular basis.
The script is open source and free to use.
Here are some additional details that you can include in the repository description:

The script requires Python 3 to run.
The script has been tested on Jellyfin 10.7.0 and Python 3.8.
The script is compatible with Windows, macOS, and Linux.
The script is well-documented and easy to use.
The script is free to use and open source.

Salient features:
The script can automatically backup watched files from your Jellyfin server to a local or remote storage device.
The script can be customized to backup specific libraries, seasons, or episodes.
The script can be scheduled to run on a regular basis.
The script is open source and free to use.

Objectives:
The objective of this script is to provide a way for Jellyfin users to automatically backup their watched files.
This can be useful for users who want to ensure that they have a backup of their watched files in case of a hard drive failure or other disaster.
The script can also be used to save space on the Jellyfin server by moving the watched files to a less expensive storage device.

Advantages:
The script is easy to use and does not require any special knowledge to set up.
The script is compatible with Windows, macOS, and Linux.
The script is open source and free to use.
The script can be customized to meet the specific needs of individual users.


Getting Started:

Configure the script by providing the Jellyfin server URL, API key, user ID, backup folder path, excluded libraries (if any), and supported file formats.
Set up the desired schedule using the powerful schedule library to automate the backup process at your preferred times.
Run the script, sit back, and relax as it takes care of the entire backup process, intelligently moving watched items once all seasons of a TV show have been completed.
Enjoy enhanced storage management, improved performance, and peace of mind knowing that your watched items are safely backed up on a secondary storage device.